add Google login/register button to register page
