+++ src/hooks/auth/useGoogleLogin.ts
@@ -1,5 +1,9 @@
 import { useRouter } from 'next/router';
 
+import {
+  GoogleLoginResponse,
+  GoogleLoginResponseOffline,
+} from 'react-google-login';
 import { useMutation } from 'react-query';
 
 import { useAxios } from '@hooks';
@@ -12,14 +16,14 @@ export const useGoogleLogin = () => {
   const router = useRouter();
   const { setIsLoggedIn } = useUserContext();
 
-  const googleLogin = async (token: string): Promise<LoginResponse> => {
+  const loginWithGoogle = async (token: string): Promise<LoginResponse> => {
     const response = await axios.post('/auth/login/google', { token });
 
     const data = response.data as LoginResponse;
     return data;
   };
 
-  return useMutation(googleLogin, {
+  const googleLogin = useMutation(loginWithGoogle, {
     onSuccess: () => {
       setIsLoggedIn(true);
 
@@ -34,4 +38,21 @@ export const useGoogleLogin = () => {
       console.log(error);
     },
   });
+
+  const onGoogleResponse = (
+    response: GoogleLoginResponse | GoogleLoginResponseOffline
+  ) => {
+    if ('accessToken' in response) {
+      googleLogin.mutate(response.accessToken);
+    }
+  };
+
+  const onGoogleFailure = (error: any) => {
+    console.log(error);
+  };
+
+  return {
+    onGoogleResponse,
+    onGoogleFailure,
+  };
 };
